requirements.txt
Flask==3.1.2
pandas==2.3.3
python-dotenv==1.1.1
Requests==2.32.5
waitress==3.0.2
